Changes.
Added the ability for the map type to be read using the config reader, which required the code in ConfigReader.java to be updated/modified and the code in MultiServer now uses the config reader to pass in the type of maze/map into the world params so you will be able to change the map without passing in command line arguments or hard coding the maze simply change the maze/map by changing the value of map in the configprop file.

Added an acceptance test for looking and finding an obstacle and modified the look and finding a robot test so that it loops if it hits an edge case where it doesn't see a robot.